# openapi.json
{"openapi":"3.1.0","info":{"title":"Wasla AI Agent APIs","summary":"Agentic AI backend for Wasla CRM and Customer Portals","description":"This API provides agentic conversational interfaces for both the Wasla Customer Portal and the Internal Company CRM. It operates natively using local Ollama models (e.g. Qwen 2.5) or configured OpenRouter models to process intent and automatically execute backend actions.\n\n## Endpoints\n\n| Endpoint | Transport | Context | Capabilities |\n|----------|-----------|---------|--------------|\n| `POST /api/chat/{company_id}` | JSON | Public Customer Portal | General inquiries, submitting service requests, checking offers. |\n| `POST /api/company/chat` | JSON | Internal CRM | Full CRM capabilities (Customers, Offers, Tasks, Appointments). Requires JWT Bearer token authentication. |\n\n---\n**Authentication:** Include standard `Bearer <JWT_TOKEN>` in the headers when accessing protected internal endpoints.","contact":{"name":"Wasla AI","url":"https://wasla.ai/"},"license":{"name":"MIT"},"version":"2.1.0"},"paths":{"/api/chat":{"post":{"tags":["Chat"],"summary":"Customer Portal — Agentic chat with tool calling","description":"**Customer Portal chat endpoint** — agentic tool-calling loop.\n\nThe LLM can invoke 27 tools covering the full Customer Portal API:\nauth, companies, reviews, profiles, offers, service requests, dashboard.\n\nClick the **lock icon** (top-right) and paste your JWT to authenticate.\nPublic actions (browse companies, view reviews) work without auth.","operationId":"portalChat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful AI response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"429":{"description":"Rate limit exceeded."},"503":{"description":"AI model is temporarily unavailable."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/company/chat":{"post":{"tags":["Company Chat"],"summary":"Company Portal — Staff agentic chat","description":"Staff-facing chat endpoint with 40 tools for CRM operations:\ncustomers, offers, tasks, employees, expenses, dashboard, service requests.\n\nClick the lock icon and paste a staff JWT to authenticate.","operationId":"companyChat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"AI response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"503":{"description":"AI unavailable."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/health":{"get":{"tags":["Health"],"summary":"Service health check","description":"Returns the service status, configured model names, and key settings.\n\nUse this endpoint to verify the API is running and to inspect which\nmodels and voices are currently active.","operationId":"health_check_health_get","responses":{"200":{"description":"Current service status and configured models.","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ChatRequest":{"properties":{"prompt":{"type":"string","minLength":1,"title":"Prompt","description":"The user's message to the AI assistant.","examples":["Show me the top 5 customers"]},"conversation_history":{"items":{"type":"object"},"type":"array","title":"Conversation History","description":"Previous messages for multi-turn context. Each dict must have `role` (user/assistant/system) and `content` keys.","examples":[[{"content":"Hello","role":"user"},{"content":"Hi! How can I help you?","role":"assistant"}]]}},"type":"object","required":["prompt"],"title":"ChatRequest","description":"Body for the main chat endpoint (Route 1).","examples":[{"conversation_history":[],"prompt":"Show me the top 5 customers"},{"conversation_history":[{"content":"Show me the top 5 customers","role":"user"},{"content":"Here are your top 5 customers...","role":"assistant"}],"prompt":"Tell me more about the first one"}]},"ChatResponse":{"properties":{"response":{"type":"string","title":"Response","description":"The AI assistant's response text (may contain Markdown)."},"tool_calls_made":{"type":"integer","title":"Tool Calls Made","description":"Number of tool calls executed during this request.","default":0},"model_used":{"type":"string","title":"Model Used","description":"The HF model that produced the response.","default":""}},"type":"object","required":["response"],"title":"ChatResponse","description":"JSON response from the main chat endpoint.","examples":[{"model_used":"meta-llama/Llama-3.3-70B-Instruct","response":"Here are your top 5 customers:\n\n1. Acme Corp — $12,400\n2. ...","tool_calls_made":1}]},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}},"tags":[{"name":"Customer Chat","description":"Customer-facing AI chat endpoint widget using an agentic tool-calling loop. Accessed externally via the Wasla Customer Portal."},{"name":"Company Chat","description":"Staff-facing (Manager / Employee) AI chat with 40 tools for full CRM operations."},{"name":"Health","description":"Service health check and configuration status."}]}